+++ internal/service/commit/stats.go
@@ -2,12 +2,21 @@ package commit
 
 import (
 	"golang.org/x/net/context"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/codes"
 
 	pb "gitlab.com/gitlab-org/gitaly-proto/go"
+	"gitlab.com/gitlab-org/gitaly/internal/rubyserver"
 )
 
 func (server) CommitStats(ctx context.Context, in *pb.CommitStatsRequest) (*pb.CommitStatsResponse, error) {
-	return nil, grpc.Errorf(codes.Unimplemented, "not implemented")
+	client, err := rubyserver.CommitServiceClient(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	clientCtx, err := rubyserver.SetHeaders(ctx, in.GetRepository())
+	if err != nil {
+		return nil, err
+	}
+
+	return client.CommitStats(clientCtx, in)
 }